ECommerce_API/requirement.txt
django
django-environ
django-extensions
djangorestframework
markdown
django-filter
mysqlclient
django-rest-framework-social-oauth2
